feat: transfer ownership api product BE changes

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/api_product/use_case/members/DeleteApiProductMemberUseCase.java
@@ -15,10 +15,17 @@
  */
 package io.gravitee.apim.core.api_product.use_case.members;
 
+import static io.gravitee.rest.api.model.permissions.SystemRole.PRIMARY_OWNER;
+
 import io.gravitee.apim.core.UseCase;
+import io.gravitee.apim.core.member.model.Member;
 import io.gravitee.apim.core.member.model.MembershipMemberType;
 import io.gravitee.apim.core.member.model.MembershipReferenceType;
 import io.gravitee.apim.core.member.query_service.MemberQueryService;
+import io.gravitee.rest.api.model.permissions.RoleScope;
+import io.gravitee.rest.api.service.exceptions.PrimaryOwnerRemovalException;
+import java.util.List;
+import java.util.Optional;
 import lombok.AllArgsConstructor;
 
 @UseCase
@@ -32,6 +39,17 @@ public class DeleteApiProductMemberUseCase {
     public record Output() {}
 
     public Output execute(Input input) {
+        Member member = memberQueryService.getUserMember(MembershipReferenceType.API_PRODUCT, input.apiProductId, input.memberId);
+        if (
+            member != null &&
+            Optional.ofNullable(member.getRoles())
+                .orElse(List.of())
+                .stream()
+                .anyMatch(r -> RoleScope.API_PRODUCT.equals(r.getScope()) && PRIMARY_OWNER.name().equals(r.getName()))
+        ) {
+            throw new PrimaryOwnerRemovalException();
+        }
+
         memberQueryService.deleteReferenceMember(
             MembershipReferenceType.API_PRODUCT,
             input.apiProductId,
